+++ src/main/java/life/qbic/subscriptions/security/SecurityConfiguration.java
@@ -32,7 +32,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
   protected void configure(HttpSecurity http) throws Exception {
     http.csrf().disable()
         .authorizeRequests()
-        .antMatchers(HttpMethod.GET, "/subscription/cancel").hasRole("OFFICER")
+        .antMatchers(HttpMethod.POST, "/subscriptions/tokens").hasRole("OFFICER")
         .and().httpBasic().realmName(REALM).authenticationEntryPoint(getBasicAuthEntryPoint())
         .and().sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);//We don't need sessions to be created.
   }
@@ -44,7 +44,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 
   /* To allow Pre-flight [OPTIONS] request from browser */
   @Override
-  public void configure(WebSecurity web) throws Exception {
+  public void configure(WebSecurity web) {
     web.ignoring().antMatchers(HttpMethod.OPTIONS, "/**");
   }